Structure change so ckb-fips205-utils becomes single truth of param id definition
