--- build-tools/build-params-finder/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "build-params-finder"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { version = "4.5.28", features = ["cargo", "derive"] }
