+++ crates/ckb-fips205-utils/Cargo.toml
@@ -1,18 +1,24 @@
 [package]
 name = "ckb-fips205-utils"
 version = "0.1.0"
+description = "A utility crate gluing CKB with FIPS 205"
 edition = "2021"
 
 [dependencies]
-ckb-std = "0.16.3"
-ckb-rust-std = "1.0.0"
-ckb-gen-types = { version = "0.119.0", default-features = false }
-ckb-hash = { version = "0.119.0", default-features = false, features = ["ckb-contract"] }
-molecule = { version = "0.8", default-features = false }
+# Verifying only crates
+ckb-std = { version = "0.16.3", optional = true }
+ckb-rust-std = { version = "1.0.0", optional = true }
+ckb-gen-types = { version = "0.119.0", default-features = false, optional = true }
+ckb-hash = { version = "0.119.0", default-features = false, features = ["ckb-contract"], optional = true }
+molecule = { version = "0.8", default-features = false, optional = true }
 
 # Signing only crates
 ckb-mock-tx-types = { version = "0.119.0", optional = true }
 rand_core = { version = "0.6.4", optional = true }
+
+# Message only crates
+sha2 = { version = "0.10.8", optional = true }
+sha3 = { version = "0.10.8", optional = true }
 
 # Serde crates
 serde = { version = "1.0", features = ["derive"], optional = true }
@@ -22,6 +28,7 @@ unicase = { version = "2.6.0", optional = true }
 [dependencies.fips205]
 version = "0.4.1"
 default-features = false
+optional = true
 features = [
   "slh_dsa_sha2_128s",
   "slh_dsa_shake_128s",
@@ -38,6 +45,15 @@ features = [
 ]
 
 [features]
-default = []
-signing = ["fips205/default-rng", "ckb-mock-tx-types", "rand_core"]
-serde = ["dep:serde", "serde_string_enum", "unicase"]
+default = ["verifying"]
+std = []
+verifying = [
+  "ckb-std", "ckb-rust-std", "ckb-gen-types",
+  "ckb-hash", "molecule", "fips205"
+]
+signing = [
+  "std", "verifying",
+  "fips205/default-rng", "ckb-mock-tx-types", "rand_core",
+]
+message = ["std", "sha2", "sha3"]
+serde = ["std", "dep:serde", "serde_string_enum", "unicase"]
