+++ crates/script-merge-tool/src/main.rs
