--- params.txt
@@ -1,12 +1,0 @@
-1 sphincs-sha2-128f
-2 sphincs-sha2-128s
-3 sphincs-sha2-192f
-4 sphincs-sha2-192s
-5 sphincs-sha2-256f
-6 sphincs-sha2-256s
-7 sphincs-shake-128f
-8 sphincs-shake-128s
-9 sphincs-shake-192f
-10 sphincs-shake-192s
-11 sphincs-shake-256f
-12 sphincs-shake-256s
