+++ tests/nist-test-vector-tests/Cargo.toml
@@ -12,10 +12,8 @@ serde_json = "1.0"
 serde_string_enum = "0.2.1"
 unicase = "2.6.0"
 
-sha2 = "0.10.8"
-sha3 = "0.10.8"
 
 [dependencies.ckb-fips205-utils]
 path = "../../crates/ckb-fips205-utils"
 default-features = false
-features = ["serde"]
+features = ["serde", "message"]
