+++ tests/sphincsplus_rust/tests/test_sphincsplus.rs
@@ -105,3 +105,16 @@ fn test_err_message() {
         panic!("pass verification");
     }
 }
+
+#[test]
+fn test_pubkey_and_signature_lengths() {
+    let param_id: ckb_fips205_utils::ParamId = ckb_sphincs_utils::sphincsplus::param_id()
+        .expect("param id must exist")
+        .try_into()
+        .expect("parse param id");
+    let (pubkey_length, signature_length) = ckb_fips205_utils::lengths(param_id);
+
+    let key = ckb_sphincs_utils::SphincsPlus::default();
+    assert_eq!(key.get_pk_len(), pubkey_length);
+    assert_eq!(key.get_sign_len(), signature_length);
+}
