+++ tools/ckb-sphincs-utils/Cargo.toml
@@ -26,5 +26,9 @@ serialize_key = ["dep:base64"]
 [dependencies]
 base64 = { version = "0.22.0", optional = true }
 
+[dependencies.ckb-fips205-utils]
+path = "../../crates/ckb-fips205-utils"
+default-features = false
+
 [build-dependencies]
 cc = "1.0.79"
